changed cli_db.py name to mopdb.py moved yaml files in data and code out of data
