--- mopper/press_levs.yaml
@@ -1,65 +1,0 @@
-################################################################
-#
-# Settings for pressure levels
-# These have been placed here from app_functions.py
-
-levels:
-    plev19: [100000, 
-             92500, 
-             85000, 
-             70000, 
-             60000, 
-             50000, 
-             40000, 
-             30000, 
-             25000, 
-             20000, 
-             15000, 
-             10000, 
-             7000, 
-             5000, 
-             3000, 
-             2000,
-             1000, 
-             500, 
-             100]
-
-    plev19min: [-1.0000e+02, 
-                3.0000e+02, 
-                7.5000e+02, 
-                1.5000e+03, 
-                2.5000e+03,
-                4.0000e+03, 
-                6.0000e+03, 
-                8.5000e+03, 
-                1.2500e+04,
-                1.7500e+04,
-                2.2500e+04, 
-                2.7500e+04, 
-                3.5000e+04, 
-                4.5000e+04, 
-                5.5000e+04,
-                6.5000e+04, 
-                7.7500e+04, 
-                8.8750e+04, 
-                9.6250e+04]
-
-    plev19max: [3.0000e+02, 
-                7.5000e+02, 
-                1.5000e+03, 
-                2.5000e+03,
-                4.0000e+03, 
-                6.0000e+03, 
-                8.5000e+03, 
-                1.2500e+04, 
-                1.7500e+04,
-                2.2500e+04, 
-                2.7500e+04, 
-                3.5000e+04, 
-                4.5000e+04, 
-                5.5000e+04,
-                6.5000e+04, 
-                7.7500e+04, 
-                8.8750e+04, 
-                9.6250e+04, 
-                1.0375e+05]
